% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_interval}
\alias{find_interval}
\title{Find the interval of a number in a sequence}
\usage{
find_interval(x, breaks)
}
\arguments{
\item{x}{A numeric value.}

\item{breaks}{A vector of numbers}
}
\description{
Find the interval of a number in a sequence
}
