% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pit_function}
\alias{pit_function}
\title{Calculate personal income tax liability based on a tax table, rebate table and medical tax credits}
\usage{
pit_function(income, age, mtc, brackets, rebates)
}
\arguments{
\item{income}{taxable income for the tax year}

\item{age}{age of each individual at the end of the tax year}

\item{mtc}{medical tax credits available in the tax year}

\item{table}{a matrix representing any tax table, specifically the marginal tax rates and brackets}

\item{rebate}{a matrix representing the rebates available for each age group}
}
\description{
Function to calculate tax liability using custom tax tables
}
\details{
\code{pit_manual} takes a vector of income and age and a matrix representation of the income tax brackets and
the rebate available per age to calculate the tax liability.
Allowing for custom tax tables make it easier to model policy changes.
Since it is implemented in C++ it should be relatively quick and well-suited to large datasets.
}
