% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARB}
\alias{SARB}
\title{South African Reserve Bank (SARB) Quarterly Bulletin data}
\format{
A data frame with 599,026 rows and 4 variables:
\describe{
  \item{Date}{Date in 8 digits, e.g. 19980101}
  \item{Frequency}{Time period of data series, e.g. J1, which is Annually - 1 January to 31 December}
  \item{Code}{Time series indicator, e.g. KBP6006J, which is Gross Domestic Product at market prices}
  \item{Value}{...}
}
}
\source{
\url{https://www.resbank.co.za/Publications/QuarterlyBulletins/Pages/DownloadInformationFromXLSXDataFiles.aspx}
}
\usage{
SARB
}
\description{
A dataset containing time series information published by the SARB in their
Quarterly Bulletin.
}
\keyword{datasets}
