% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARB_descriptions}
\alias{SARB_descriptions}
\title{Descriptions of South African Reserve Bank (SARB) Quarterly Bulletin variables}
\format{
A data frame with 4,047 rows and 6 variables:
\describe{
  \item{Code}{Code for specific time series, e.g. KBP6006J1}
  \item{Description}{Description of time series, e.g. Gross domestic product at market prices}
  \item{Frequency}{Time period of data series, e.g. J1 }
  \item{Frequency_description}{Description of time period, e.g. Annually - 1 January to 31 December}
  \item{Unit_of_measure}{Characteristic of the value, e.g. RMILL (R millions)}
  \item{Version_Description}{Additional information on value}
}
}
\source{
\url{https://www.resbank.co.za/Publications/QuarterlyBulletins/Pages/ListOfDesciptionsOfAllQuarterlyBulletinTimeSeries.aspx}
}
\usage{
SARB_descriptions
}
\description{
A dataset containing information on the variables that are published by the SARB in their
Quarterly Bulletin.
}
\keyword{datasets}
