% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARS}
\alias{SARS}
\title{Monthly tax revenue data from the South African Revenue Service}
\format{
A data frame with 14,764 rows and 5 variables:
\describe{
  \item{Tax}{Type of tax instrument}
  \item{Year}{Calendar year}
  \item{Fiscal_year}{Year running from April 1 to March 31}
  \item{Month}{}
  \item{Revenue}{Tax revenue in R thousands}
}
}
\source{
\url{http://www.treasury.gov.za/comm_media/press/monthly/default.aspx}
}
\usage{
SARS
}
\description{
A dataset containing tax revenue by tax instrument
per month that is collected by the South African Revenue Service (SARS) and
published by the National Treasury of South Africa.
}
\keyword{datasets}
