% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STATSSA}
\alias{STATSSA}
\title{Data from Statistics South Africa}
\format{
A data frame with 703,779 rows and 4 variables:
\describe{
  \item{Publication}{Publication or report number}
  \item{Code}{H03 from publication}
  \item{Date}{Monthly, quarterly or annual date}
  \item{Value}{}
}
}
\source{
\url{http://www.statssa.gov.za/?page_id=1847}
}
\usage{
STATSSA
}
\description{
Covers a variety of variables, from inflation to electricity generation
}
\keyword{datasets}
