% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STATSSA_descriptions}
\alias{STATSSA_descriptions}
\title{Descriptions of Statistics South Africa (STATSSA) variables}
\format{
A data frame with 3,495 rows and 17 variables:
\describe{
  \item{H01}{Publication or report}
  \item{...}{...}
  \item{H25}{Time period}
}
}
\source{
\url{http://www.statssa.gov.za/?page_id=1847}
}
\usage{
STATSSA_descriptions
}
\description{
A dataset containing information on the variables that are published by the STATSSA
}
\keyword{datasets}
