% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_calculation.R
\name{tax_calculation}
\alias{tax_calculation}
\title{Calculate tax liability based on a tax table}
\usage{
tax_calculation(income, tax_table)
}
\arguments{
\item{income}{any amount to apply to a tax table}

\item{tax_table}{a matrix representing any tax table}
}
\description{
Function to calculate tax liability using custom tax tables
}
\details{
\code{tax_calculation} takes a vector of income and a matrix representation of the income tax brackets to calculate the tax liability.
Allowing for custom tax tables make it easier to model policy changes.
}
