% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tax_tables}
\alias{Tax_tables}
\title{South African tax tables from 2010/11 to 2020/21}
\format{
A list with 22 matrices:
\describe{
  \item{PIT_brackets_year}{The tax brackets and marginal tax rates}
  \item{PIT_rebates_year}{The primary, secondary and tertiary rebates}
}
}
\usage{
Tax_tables
}
\description{
A list containing the tax brackets and tax rebates to calculate the personal income tax liability in a given year.
}
\keyword{datasets}
